move RevaultDError
