+++ src/app/error.rs
@@ -1,4 +1,4 @@
-use crate::daemon::client::RevaultDError;
+use crate::daemon::RevaultDError;
 use revaultd::config::ConfigError;
 use std::convert::From;
 use std::io::ErrorKind;
